Closes issue #50.
I changed the global AppState to include the RequestConfig-object
instead of a Request -> Request function. The problem was that
its"createReferer"-field could not simply be turned into a function of
the form Request -> Request. Now the entire configuration object is
passed into the CLI.

The static arguments (Library.StaticArgs) now has an additional,
Bool-type parameter, which is filled in by the CLI, from the
RequestConfig-object. This Bool-type argument, which indicates whether a
HTTP referer ought to be created, is passed around basically everywhere
until it reaches Fetch.Types.Successor.htmlSuccessor, which is the
function that actually adds the HTTP referer.

This also means that the "createReferer"-setting is ONLY effective if
the successor functions for crawlers are created via htmlSuccessor. If
another way is used, the user is on his own.