Closes #72.
I removed the nodeURL field from SuccessorNode type and instead put a field of type URI (not URL!) into Inner and Blob. This had wide-reaching consequences, as URIs have to be parsed as such, whereas Text does not have to be. As a result, I pushed to parsing out into Crawlers.Library, where it is done before the any crawling even starts.