Added identifiers to FetchResults.
Relates to issue #77.

Also, I re-arranged the type variables of FetchResult, SuccessorNode, etc. - it's now 'e i a' instead of 'i e a'. First, the exception, then the identifier type, then the node state (if applicable). In Crawlers.Library, i is converted into an existential 'Ident' type, which has the constraints (Show, ToJSON).