Added printDownloadSummary, removed the mutex thing CLI.Status.
A CLI mutex, as it is now implemented, is really not workable because
a function that holds it can't grant it to another one it calls
(in the same thread) without a deadlock between caller and callee.
As it currently stands, only the CLI's main thread is doing output, so
the lack of synchronisation on stdout is not a problem, but this issue should
be revisited at a later date (perhaps via having mutexes be of type TVar (ThreadId)).